Fixing a compile-time field assignment bug
The bug was when we had a object constructed like so:
class C1
  let f: U32
  new create(f': U32) => f = f'

class C2
  let c: C1
  new create(f: U32) => C1(f)

actor Main
  new create(env: Env) =>
    let c = # C2(2)

The object C1 object constructed during the construction of C2 did not have its
field f set. Instead what happened is that the value setting would first look
for an existing definition of 'f', find one (namely the parameter to 'create')
and set that instead. This had the effect that if we tried to use the object
elsewhere (outside of the C2 scope) then it's field would be missing a value.

This commit fixes that by allowing value settings both looking for an existing
value or now. Also provides a test case.